requirements.txt
# Podstawowe dla MVP
streamlit==1.31.1
supabase==2.3.1
openai==1.12.0
python-dotenv==1.0.0
plotly==5.18.0

# Przetwarzanie danych
pandas==2.1.4
numpy==1.26.3

# Testowanie
pytest==7.4.3
pytest-cov==4.1.0

# Formatowanie i linting
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Bezpieczeństwo
python-jose==3.3.0
passlib==1.7.4